Add initial content and style to help pages.
A first attempt was made to fill in the bodies of the help pages and
create a pleasing style. Content is sparse, but will be built on once
the site in general has more content to feature. The Getting Started
page, in particular, has no content because it requires screen-shots
and a site map that can't be made available yet.

    - Wrote preliminary copy.
    - Added content to contact.html.haml, faq.html.haml,
      getting_started.html.haml, and help_center.html.haml.
    - Added general, minimal styling.
    $ git add .
    $ git commit -F commit-message.txt
    $ git push origin master